Perform AffineTransformation compuation on doubles
We lose too much precision when running on floats.

+++ internal/core/regularsurface.cpp
@@ -12,11 +12,11 @@ Point AffineTransformation::operator*(Point p) const noexcept (true) {
 
 
 AffineTransformation AffineTransformation::from_rotation(
-    float xori,
-    float yori,
-    float xinc,
-    float yinc,
-    float rot
+    double xori,
+    double yori,
+    double xinc,
+    double yinc,
+    double rot
 ) noexcept (true) {
     double rad = rot * (M_PI / 180);
     /**
